+++ R/geoset.R
@@ -10,20 +10,27 @@
 #' Set \code{tidy = FALSE} to return only the initial result of \code{jsonlite::fromJSON}.
 #' By default, additional processing is done to return a tibble data frame.
 #'
-#' @param api_key character, API key.
+#' Set to \code{cache = FALSE} to force a new API call for updated data.
+#' Using \code{FALSE} always makes a new API call and returns the result from the server.
+#' \code{TRUE} uses memoization on a per R session basis, caching the result of the function call in memory for the duration of the R session.
+#' You can reset the entire cache by calling \code{eia_clear_cache()}.
+#'
+#' @param key character, API key.
 #' @param id character, geoset series ID, may be a vector. See details.
 #' @param region character, region ID, may be a vector. Data available for the intersection of \code{id} and \code{region} is returned.
 #' @param start start date. Providing only a start date will return up to the maximum 100 results if available.
 #' @param end end date. Providing only an end date will a single result for that date.
 #' @param n integer, length of series to return ending at most recent value or at \code{end} date if also provided. Ignored if \code{start} is not \code{NULL}.
 #' @param tidy logical, return a tidier result. See details.
+#' @param cache logical, cache result for duration of R session using memoization. See details.
 #'
 #' @return a tibble data frame
 #' @export
+#' @seealso \code{\link{eia_clear_cache}}
 #'
 #' @examples
 #' \dontrun{
-#' key <- readRDS("key.rds") # your stored API key
+#' key <- Sys.getenv("EIA_KEY") # your stored API key
 #' id <- paste0("ELEC.GEN.ALL-99.", c("A", "Q", "M"))
 #' region <- c("USA-CA", "USA-NY")
 #'
@@ -35,13 +42,21 @@
 #' x <- eia_geoset(key, id, region[1], end = 2016, n = 10)
 #' x$data[[1]]
 #' }
-eia_geoset <- function(api_key, id, region, start = NULL, end = NULL, n = NULL, tidy = TRUE){
-  f <- if(tidy) purrr::map_dfr else purrr::map
-  f(id, ~.eia_geoset_by_id(api_key, .x, region, start, end, n, tidy))
+eia_geoset <- function(key, id, region, start = NULL, end = NULL, n = NULL,
+                       tidy = TRUE, cache = TRUE){
+  if(cache) .eia_geoset_memoized(key, id, region, start, end, n, tidy) else
+    .eia_geoset(key, id, region, start, end, n, tidy)
 }
 
-.eia_geoset_by_id <- function(api_key, id, region, start = NULL, end = NULL, n = NULL, tidy = TRUE){
-  x <- .eia_geo_url(api_key, id, region, start, end, n) %>% httr::GET() %>%
+.eia_geoset <- function(key, id, region, start = NULL, end = NULL, n = NULL, tidy = TRUE){
+  f <- if(tidy) purrr::map_dfr else purrr::map
+  f(id, ~.eia_geoset_by_id(key, .x, region, start, end, n, tidy))
+}
+
+.eia_geoset_memoized <- memoise::memoise(.eia_geoset)
+
+.eia_geoset_by_id <- function(key, id, region, start = NULL, end = NULL, n = NULL, tidy = TRUE){
+  x <- .eia_geo_url(key, id, region, start, end, n) %>% httr::GET() %>%
     httr::content(as = "text", encoding = "UTF-8") %>%
     jsonlite::fromJSON()
   if(!tidy) return(x)
@@ -69,9 +84,9 @@ eia_geoset <- function(api_key, id, region, start = NULL, end = NULL, n = NULL, 
   dplyr::bind_cols(x$geoset, x$series)
 }
 
-.eia_geo_url <- function(api_key, id, region, start = NULL, end = NULL, n = NULL){
+.eia_geo_url <- function(key, id, region, start = NULL, end = NULL, n = NULL){
   params <- .eia_time_params(start, end, n, n_default = 1)
-  url <- .eia_url(api_key, id, "geoset")
+  url <- .eia_url(key, id, "geoset")
   url <- paste0(url, "&regions=", paste0(region, collapse = ";"))
   if(!is.null(params$start)) url <- paste0(url, "&start=", params$start)
   if(!is.null(params$end)) url <- paste0(url, "&end=", params$end)