+++ R/zzz.R
@@ -1,0 +1,5 @@
+.onLoad <- function(libname, pkgname){
+  options(eia_antidos = 1)
+}
+
+eia_api_time <- new.env()