+++ NEWS.md
@@ -1,3 +1,11 @@
+# eia 0.2.0
+
+* Added optional memoization to API functions, adding a new `cache` argument.
+* Added anti-DOS measures, which can be adjusted using `options()`.
+* Added more documentation.
+* Added helper functions for clearing cached results.
+* Minor updates to functions, documentation and unit test scripts.
+
 # eia 0.1.0
 
 * Added package scaffolding.